requirements.txt
asgiref==3.2.10
Django==3.1
django-crontab==0.7.1
djangorestframework==3.11.1
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
